Binary Search neetcode
You are given an array of distinct integers nums, sorted in ascending order, and an integer target.

Implement a function to search for target within nums. If it exists, then return its index, otherwise, return -1.

Your solution must run in 
O(logn) time.

What is Binary search? It's when you divide the problem set into half and keep doing so until you find the correct answer.

Analogy: Imagine you’re in a tall building with an egg, and you need to figure out the highest floor from which you can drop the egg without it breaking. Instead of testing every floor one by one, you use a smarter strategy: you divide the building into sections and narrow down the possibilities. For example, you might first ask, "Is the safe floor between 1 and 50?" If the answer is no, you then ask, "Is it between 50 and 75?" If the answer is yes, you’ve narrowed it down to a smaller range, making it easier to find the exact floor. This method is similar to how binary search works—it efficiently narrows down the search space by repeatedly dividing it in half.